Added ability to get cheapest value when retrieving rates from FedEx
Current logic for DotNetShipping uses the Max value of the returned rates for a particular provider code. This becomes a problem when using the LIST RateRequestType since LIST will return discounted rates and non-discounted rates. If you want to present discounted rates, you're not able to do such. Now, using the FedExRetrieveCheapestRates config value (or parameter) and setting it to true, it'll instead get the Min value for the total net charge.